Improved values system
### Utils
- deleted `Tokenizer`

- created `Structs`
  at `org.jamplate.glucose.internal.util`
- created `Values`
  at `org.jamplate.glucose.internal.util`

### Values
- renamed `GluedValue` to `GlueValue`

### Instructions
- deleted `JoinFrame`

- created `GlueFrame`
  at `org.jamplate.glucose.instruction.memory.frame`

- created `BuildArray`
  at `org.jamplate.glucose.instruction.operator.cast`
- created `BuildObject`
  at `org.jamplate.glucose.instruction.operator.cast`
- created `BuildPair`
  at `org.jamplate.glucose.instruction.operator.cast`

- created `CastGlue`
  at `org.jamplate.glucose.instruction.operator.cast`

- optimized instruction `Get`
- optimized instruction `Put`
- optimized instruction `Split`
- optimized instruction `Touch`

### Specs
- optimized `RootSpec`
- optimized `TextSpec`
- optimized `TextSpec`
- optimized `InjectionSpec`
- optimized `GetterSpec`
- optimized `PairSpec`
- optimized `ArraySpec`
- optimized `ObjectSpec`

### Others
- general optimizations
- updated tests

+++ src/main/java/org/jamplate/glucose/instruction/operator/math/Difference.java
@@ -105,7 +105,8 @@ public class Difference implements Instruction {
 
 			//result
 			NumberValue number2 = number1.apply((m, n) ->
-					n - number0.getPipe().eval(m)
+					n.doubleValue() -
+					number0.getPipe().eval(m).doubleValue()
 			);
 
 			memory.push(number2);
